requirements.txt
agate==1.7.1
alembic==1.12.1
attrs==23.1.0
Babel==2.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
databricks-sdk==0.13.0
databricks-sql-connector==2.9.3
dbt-core==1.6.9
dbt-databricks==1.6.6
dbt-extractor==0.4.1
dbt-semantic-interfaces==0.2.3
dbt-spark==1.6.0
et-xmlfile==1.1.0
hologram==0.0.16
idna==3.6
importlib-metadata==6.8.0
isodate==0.6.1
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keyring==24.3.0
leather==0.3.4
Logbook==1.5.3
lz4==4.3.2
Mako==1.3.0
MarkupSafe==2.1.3
mashumaro==3.8.1
minimal-snowplow-tracker==0.0.2
more-itertools==8.14.0
msgpack==1.0.7
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
parsedatetime==2.6
pathspec==0.11.2
protobuf==4.25.1
pyarrow==14.0.1
pycparser==2.21
pydantic==1.10.13
python-dateutil==2.8.2
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
six==1.16.0
SQLAlchemy==1.4.50
sqlparams==5.1.0
sqlparse==0.4.4
text-unidecode==1.3
thrift==0.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
zipp==3.17.0
